Reclaim mechanism to get back oprhaned blocks.
The idea is to just erase any blocks that aren't taken to ensure the
allocator can find them again. This is to ensure failures while
writing don't orphan blocks on accidentally, since the allocator looks
for erased blocks. So if a file is being written and crashes before
finishing we could lose those blocks.