A test that reproduces one of the corruption issues I've seen!
See FK-594 for more details. Basic gist is this:

We're going along, write a large block that spans sectors 14 and 15 of a block.
Then a sudden restart happens and when the file is opened again, we resume
writing on block 15. This blows away the "tail" of the record that spanned the
blocks. This only seems to happen when we're spanning into the last sector.