Workaround for a strange seeking issue.
I think there's a bug somewhere that causes index entries to be skipped and so
the file grew so large the linked block searching ended up taking too long. This
will write index blocks as we go if we find them to be necessary so we can fix
the situation if it occurs.